Login,Register UserController,...

+++ src/main/java/com/project/NovelWeb/dtos/UserDTO.java
@@ -27,5 +27,6 @@ public class UserDTO {
 
     @NotNull(message = "Role Id is required!")
     @JsonProperty("role_id")
-    private Integer roleId;
+    private Long roleId;
+
 }
